+++ scripts/tests/test_volume_reslice_to_reference.py
@@ -13,7 +13,7 @@ tmp_dir = tempfile.TemporaryDirectory()
 
 
 def test_help_option(script_runner):
-    ret = script_runner.run('scil_volume_reshape_to_reference.py', '--help')
+    ret = script_runner.run('scil_volume_reslice_to_reference.py', '--help')
     assert ret.success
 
 
@@ -21,8 +21,8 @@ def test_execution_others(script_runner, monkeypatch):
     monkeypatch.chdir(os.path.expanduser(tmp_dir.name))
     in_img = os.path.join(SCILPY_HOME, 'others', 't1_crop.nii.gz')
     in_ref = os.path.join(SCILPY_HOME, 'others', 't1.nii.gz')
-    ret = script_runner.run('scil_volume_reshape_to_reference.py', in_img,
-                            in_ref, 't1_reshape.nii.gz',
+    ret = script_runner.run('scil_volume_reslice_to_reference.py', in_img,
+                            in_ref, 't1_reslice.nii.gz',
                             '--interpolation', 'nearest')
     assert ret.success
 
@@ -31,7 +31,7 @@ def test_execution_4D(script_runner, monkeypatch):
     monkeypatch.chdir(os.path.expanduser(tmp_dir.name))
     in_img = os.path.join(SCILPY_HOME, 'commit_amico', 'dwi.nii.gz')
     in_ref = os.path.join(SCILPY_HOME, 'others', 't1.nii.gz')
-    ret = script_runner.run('scil_volume_reshape_to_reference.py', in_img,
-                            in_ref, 'dwi_reshape.nii.gz',
+    ret = script_runner.run('scil_volume_reslice_to_reference.py', in_img,
+                            in_ref, 'dwi_reslice.nii.gz',
                             '--interpolation', 'nearest')
     assert ret.success
